Move account schemas to dedicated file
Refactored account-related Zod schemas and types from zuoraSchemas.ts to a new file, account.ts, under types/objects. This improves code organization and separation of concerns for Zuora account data structures.
