+++ modules/zuora/src/zuoraSchemas.ts
@@ -1,6 +1,5 @@
 import { BillingPeriodValues } from '@modules/billingPeriod';
 import { z } from 'zod';
-import { zuoraResponseSchema } from './types';
 
 // --------------- Auth ---------------
 export type OAuthClientCredentials = z.infer<
@@ -74,50 +73,6 @@ export type ZuoraSubscription = z.infer<typeof zuoraSubscriptionSchema>;
 export type RatePlan = ZuoraSubscription['ratePlans'][number];
 
 export type RatePlanCharge = RatePlan['ratePlanCharges'][number];
-
-// --------------- Account ---------------
-export const zuoraAccountBasicInfoSchema = z
-	.object({
-		id: z.string(),
-		IdentityId__c: z.string(),
-	})
-	.transform((obj) => ({
-		id: obj.id,
-		identityId: obj.IdentityId__c,
-	}));
-
-export const metricsSchema = z.object({
-	totalInvoiceBalance: z.number(),
-	currency: z.string(),
-});
-export const billToContactSchema = z.object({
-	firstName: z.string(),
-	lastName: z.string(),
-	workEmail: z.string(),
-});
-
-export const billingAndPaymentSchema = z.object({
-	currency: z.string(),
-	defaultPaymentMethodId: z.string(),
-});
-
-export const zuoraAccountSchema = z.object({
-	success: z.boolean(),
-	basicInfo: zuoraAccountBasicInfoSchema,
-	billingAndPayment: billingAndPaymentSchema,
-	billToContact: billToContactSchema,
-	metrics: metricsSchema,
-});
-
-export type ZuoraAccount = z.infer<typeof zuoraAccountSchema>;
-
-export const zuoraSubscriptionsFromAccountSchema = zuoraResponseSchema.extend({
-	subscriptions: z.array(zuoraSubscriptionSchema).optional(),
-});
-
-export type ZuoraSubscriptionsFromAccountResponse = z.infer<
-	typeof zuoraSubscriptionsFromAccountSchema
->;
 
 // --------------- Subscribe ---------------
 export const zuoraSubscribeResponseSchema = z.array(
